extract lookup join base
