+++ src/batch/src/executor/join/distributed_lookup_join.rs
@@ -38,11 +38,7 @@ use uuid::Uuid;
 
 use crate::executor::join::chunked_data::ChunkedData;
 use crate::executor::join::JoinType;
-use crate::executor::{
-    utils, BoxedDataChunkListStream, BoxedDataChunkStream, BoxedExecutor, BoxedExecutorBuilder,
-    BoxedLookupExecutorBuilder, BufferChunkExecutor, DummyExecutor, EquiJoinParams, Executor,
-    ExecutorBuilder, HashJoinExecutor, JoinHashMap, LookupExecutorBuilder, RowId,
-};
+use crate::executor::{utils, BoxedDataChunkListStream, BoxedDataChunkStream, BoxedExecutor, BoxedExecutorBuilder, BoxedLookupExecutorBuilder, BufferChunkExecutor, DummyExecutor, EquiJoinParams, Executor, ExecutorBuilder, HashJoinExecutor, JoinHashMap, LookupExecutorBuilder, RowId, LookupJoinBase};
 use crate::task::{BatchTaskContext, TaskId};
 
 /// Inner side executor builder for the `DistributedLookupJoinExecutor`
@@ -173,20 +169,7 @@ impl<C: BatchTaskContext> LookupExecutorBuilder for InnerSideExecutorBuilder<C> 
 /// Distributed lookup join already scheduled to its inner side corresponding compute node, so that
 /// it can just lookup the compute node locally without sending RPCs to other compute nodes.
 pub struct DistributedLookupJoinExecutor<K> {
-    join_type: JoinType,
-    condition: Option<BoxedExpression>,
-    outer_side_input: BoxedExecutor,
-    outer_side_data_types: Vec<DataType>, // Data types of all columns of outer side table
-    outer_side_key_idxs: Vec<usize>,
-    inner_side_builder: BoxedLookupExecutorBuilder,
-    inner_side_key_types: Vec<DataType>, // Data types only of key columns of inner side table
-    inner_side_key_idxs: Vec<usize>,
-    null_safe: Vec<bool>,
-    chunk_builder: DataChunkBuilder,
-    schema: Schema,
-    output_indices: Vec<usize>,
-    identity: String,
-    chunk_size: usize,
+    base: LookupJoinBase<K>,
     _phantom: PhantomData<K>,
 }
 
@@ -194,205 +177,26 @@ const AT_LEAST_OUTER_SIDE_ROWS: usize = 512;
 
 impl<K: HashKey> Executor for DistributedLookupJoinExecutor<K> {
     fn schema(&self) -> &Schema {
-        &self.schema
+        &self.base.schema
     }
 
     fn identity(&self) -> &str {
-        &self.identity
+        &self.base.identity
     }
 
     fn execute(self: Box<Self>) -> BoxedDataChunkStream {
-        self.do_execute()
+        Box::new(self.base).do_execute()
     }
 }
 
 impl<K> DistributedLookupJoinExecutor<K> {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
-        join_type: JoinType,
-        condition: Option<BoxedExpression>,
-        outer_side_input: BoxedExecutor,
-        outer_side_data_types: Vec<DataType>,
-        outer_side_key_idxs: Vec<usize>,
-        inner_side_builder: Box<dyn LookupExecutorBuilder>,
-        inner_side_key_types: Vec<DataType>,
-        inner_side_key_idxs: Vec<usize>,
-        null_safe: Vec<bool>,
-        chunk_builder: DataChunkBuilder,
-        schema: Schema,
-        output_indices: Vec<usize>,
-        identity: String,
-        chunk_size: usize,
+        base: LookupJoinBase<K>
     ) -> Self {
         Self {
-            join_type,
-            condition,
-            outer_side_input,
-            outer_side_data_types,
-            outer_side_key_idxs,
-            inner_side_builder,
-            inner_side_key_types,
-            inner_side_key_idxs,
-            null_safe,
-            chunk_builder,
-            schema,
-            output_indices,
-            identity,
-            chunk_size,
+            base,
             _phantom: PhantomData,
-        }
-    }
-}
-
-impl<K: HashKey> DistributedLookupJoinExecutor<K> {
-    #[try_stream(boxed, ok = DataChunk, error = RwError)]
-    async fn do_execute(mut self: Box<Self>) {
-        let outer_side_schema = self.outer_side_input.schema().clone();
-
-        let null_matched = {
-            let mut null_matched = FixedBitSet::with_capacity(self.null_safe.len());
-            for (idx, col_null_matched) in self.null_safe.iter().copied().enumerate() {
-                null_matched.set(idx, col_null_matched);
-            }
-            null_matched
-        };
-
-        let mut outer_side_batch_read_stream: BoxedDataChunkListStream =
-            utils::batch_read(self.outer_side_input.execute(), AT_LEAST_OUTER_SIDE_ROWS);
-
-        while let Some(chunk_list) = outer_side_batch_read_stream.next().await {
-            let chunk_list = chunk_list?;
-
-            // Group rows with the same key datums together
-            let groups = chunk_list
-                .iter()
-                .flat_map(|chunk| {
-                    chunk.rows().map(|row| {
-                        self.outer_side_key_idxs
-                            .iter()
-                            .map(|&idx| row.value_at(idx).to_owned_datum())
-                            .collect_vec()
-                    })
-                })
-                .sorted()
-                .dedup()
-                .collect_vec();
-
-            self.inner_side_builder.reset();
-            for row_key in &groups {
-                self.inner_side_builder.add_scan_range(row_key)?;
-            }
-            let inner_side_input = self.inner_side_builder.build_executor().await?;
-
-            // Lookup join outer side will become the probe side of hash join,
-            // while its inner side will become the build side of hash join.
-            let hash_join_probe_side_input = Box::new(BufferChunkExecutor::new(
-                outer_side_schema.clone(),
-                chunk_list,
-            ));
-            let hash_join_build_side_input = inner_side_input;
-            let hash_join_probe_data_types = self.outer_side_data_types.clone();
-            let hash_join_build_data_types = hash_join_build_side_input.schema().data_types();
-            let hash_join_probe_side_key_idxs = self.outer_side_key_idxs.clone();
-            let hash_join_build_side_key_idxs = self.inner_side_key_idxs.clone();
-
-            let full_data_types = [
-                hash_join_probe_data_types.clone(),
-                hash_join_build_data_types.clone(),
-            ]
-            .concat();
-
-            let mut build_side = Vec::new();
-            let mut build_row_count = 0;
-            #[for_await]
-            for build_chunk in hash_join_build_side_input.execute() {
-                let build_chunk = build_chunk?;
-                if build_chunk.cardinality() > 0 {
-                    build_row_count += build_chunk.cardinality();
-                    build_side.push(build_chunk.compact())
-                }
-            }
-            let mut hash_map =
-                JoinHashMap::with_capacity_and_hasher(build_row_count, PrecomputedBuildHasher);
-            let mut next_build_row_with_same_key =
-                ChunkedData::with_chunk_sizes(build_side.iter().map(|c| c.capacity()))?;
-
-            // Build hash map
-            for (build_chunk_id, build_chunk) in build_side.iter().enumerate() {
-                let build_keys = K::build(&hash_join_build_side_key_idxs, build_chunk)?;
-
-                for (build_row_id, build_key) in build_keys.into_iter().enumerate() {
-                    // Only insert key to hash map if it is consistent with the null safe
-                    // restriction.
-                    if build_key.null_bitmap().is_subset(&null_matched) {
-                        let row_id = RowId::new(build_chunk_id, build_row_id);
-                        next_build_row_with_same_key[row_id] = hash_map.insert(build_key, row_id);
-                    }
-                }
-            }
-
-            let params = EquiJoinParams::new(
-                hash_join_probe_side_input,
-                hash_join_probe_data_types,
-                hash_join_probe_side_key_idxs,
-                build_side,
-                hash_join_build_data_types,
-                full_data_types,
-                hash_map,
-                next_build_row_with_same_key,
-                self.chunk_size,
-            );
-
-            if let Some(cond) = self.condition.as_ref() {
-                let stream = match self.join_type {
-                    JoinType::Inner => {
-                        HashJoinExecutor::do_inner_join_with_non_equi_condition(params, cond)
-                    }
-                    JoinType::LeftOuter => {
-                        HashJoinExecutor::do_left_outer_join_with_non_equi_condition(params, cond)
-                    }
-                    JoinType::LeftSemi => {
-                        HashJoinExecutor::do_left_semi_join_with_non_equi_condition(params, cond)
-                    }
-                    JoinType::LeftAnti => {
-                        HashJoinExecutor::do_left_anti_join_with_non_equi_condition(params, cond)
-                    }
-                    JoinType::RightOuter
-                    | JoinType::RightSemi
-                    | JoinType::RightAnti
-                    | JoinType::FullOuter => unimplemented!(),
-                };
-                // For non-equi join, we need an output chunk builder to align the output chunks.
-                let mut output_chunk_builder =
-                    DataChunkBuilder::new(self.schema.data_types(), self.chunk_size);
-                #[for_await]
-                for chunk in stream {
-                    #[for_await]
-                    for output_chunk in output_chunk_builder
-                        .trunc_data_chunk(chunk?.reorder_columns(&self.output_indices))
-                    {
-                        yield output_chunk
-                    }
-                }
-                if let Some(output_chunk) = output_chunk_builder.consume_all() {
-                    yield output_chunk
-                }
-            } else {
-                let stream = match self.join_type {
-                    JoinType::Inner => HashJoinExecutor::do_inner_join(params),
-                    JoinType::LeftOuter => HashJoinExecutor::do_left_outer_join(params),
-                    JoinType::LeftSemi => HashJoinExecutor::do_left_semi_anti_join::<false>(params),
-                    JoinType::LeftAnti => HashJoinExecutor::do_left_semi_anti_join::<true>(params),
-                    JoinType::RightOuter
-                    | JoinType::RightSemi
-                    | JoinType::RightAnti
-                    | JoinType::FullOuter => unimplemented!(),
-                };
-                #[for_await]
-                for chunk in stream {
-                    yield chunk?.reorder_columns(&self.output_indices)
-                }
-            }
         }
     }
 }
@@ -544,20 +348,23 @@ impl HashKeyDispatcher for DistributedLookupJoinExecutorArgs {
 
     fn dispatch_impl<K: HashKey>(self) -> Self::Output {
         Box::new(DistributedLookupJoinExecutor::<K>::new(
-            self.join_type,
-            self.condition,
-            self.outer_side_input,
-            self.outer_side_data_types,
-            self.outer_side_key_idxs,
-            self.inner_side_builder,
-            self.inner_side_key_types,
-            self.inner_side_key_idxs,
-            self.null_safe,
-            self.chunk_builder,
-            self.schema,
-            self.output_indices,
-            self.identity,
-            self.chunk_size,
+            LookupJoinBase::<K> {
+                join_type: self.join_type,
+                condition: self.condition,
+                outer_side_input: self.outer_side_input,
+                outer_side_data_types: self.outer_side_data_types,
+                outer_side_key_idxs: self.outer_side_key_idxs,
+                inner_side_builder: self.inner_side_builder,
+                inner_side_key_types: self.inner_side_key_types,
+                inner_side_key_idxs: self.inner_side_key_idxs,
+                null_safe: self.null_safe,
+                chunk_builder: self.chunk_builder,
+                schema: self.schema,
+                output_indices: self.output_indices,
+                chunk_size: self.chunk_size,
+                identity: self.identity,
+                _phantom: PhantomData,
+            }
         ))
     }
