+++ src/batch/src/executor/join/mod.rs
@@ -18,9 +18,11 @@ pub mod hash_join;
 pub mod lookup_join;
 pub mod nested_loop_join;
 mod sort_merge_join;
+mod lookup_join_base;
 
 pub use chunked_data::*;
 pub use distributed_lookup_join::*;
+pub use lookup_join_base::*;
 pub use hash_join::*;
 use itertools::Itertools;
 pub use lookup_join::*;
